update to v2.1.4 & local fallback for jquery
Added a local fallback for jquery so the startpage will still work in
the unlikely case of no internet connection and no cached version of
jquery being present(and in regions where google urls are blocked).
Also updated the jquery version used to 2.1.4, because people still
using IE8 or lower can suck it.